requirements.txt
streamlit>=1.0
Pillow>=9.0
